# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Pectus Aggregator",
    "contact": {
      "name": "Open Pectus",
      "url": "https://github.com/Open-Pectus/Open-Pectus"
    },
    "version": "0.1.dev"
  },
  "paths": {
    "/api/process_unit/{unit_id}": {
      "get": {
        "tags": [
          "process_unit"
        ],
        "summary": "Get Unit",
        "operationId": "get_unit",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessUnit"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/process_units": {
      "get": {
        "tags": [
          "process_unit"
        ],
        "summary": "Get Units",
        "operationId": "get_units",
        "parameters": [
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessUnit"
                  },
                  "title": "Response Get Units"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/process_unit/{engine_id}/process_values": {
      "get": {
        "tags": [
          "process_unit"
        ],
        "summary": "Get Process Values",
        "operationId": "get_process_values",
        "parameters": [
          {
            "name": "engine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engine Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessValue"
                  },
                  "title": "Response Get Process Values"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/process_unit/{engine_id}/all_process_values": {
      "get": {
        "tags": [
          "process_unit"
        ],
        "summary": "Get All Process Values",
        "operationId": "get_all_process_values",
        "parameters": [
          {
            "name": "engine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Engine Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessValue"
                  },
                  "title": "Response Get All Process Values"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/process_unit/{unit_id}/execute_command": {
      "post": {
        "tags": [
          "process_unit"
        ],
        "summary": "Execute Command",
        "operationId": "execute_command",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutableCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/process_unit/{unit_id}/process_diagram": {
      "get": {
        "tags": [
          "process_unit"
        ],
        "summary": "Get Process Diagram",
        "operationId": "get_process_diagram",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessDiagram"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/process_unit/{unit_id}/command_examples": {
      "get": {
        "tags": [
          "process_unit"
        ],
        "summary": "Get Command Examples",
        "operationId": "get_command_examples",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommandExample"
                  },
                  "title": "Response Get Command Examples"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/process_unit/{unit_id}/run_log": {
      "get": {
        "tags": [
          "process_unit"
        ],
        "summary": "Get Run Log",
        "operationId": "get_run_log",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunLog"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/process_unit/{unit_id}/method-and-state": {
      "get": {
        "tags": [
          "process_unit"
        ],
        "summary": "Get Method And State",
        "operationId": "get_method_and_state",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MethodAndState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/process_unit/{unit_id}/method": {
      "get": {
        "tags": [
          "process_unit"
        ],
        "summary": "Get Method",
        "operationId": "get_method",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Method"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "process_unit"
        ],
        "summary": "Save Method",
        "operationId": "save_method",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Method"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MethodVersion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/process_unit/{unit_id}/plot_configuration": {
      "get": {
        "tags": [
          "process_unit"
        ],
        "summary": "Get Plot Configuration",
        "operationId": "get_plot_configuration",
        "parameters": [
          {
            "name": "unit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Unit Id"
            }
          },
          {
            "name": "x-identity",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Identity"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlotConfiguration"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandExample": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "example": {
            "type": "string",
            "title": "Example"
          }
        },
        "type": "object",
        "required": [
          "name",
          "example"
        ],
        "title": "CommandExample"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProcessDiagram": {
        "properties": {
          "svg": {
            "type": "string",
            "title": "Svg"
          }
        },
        "type": "object",
        "required": [
          "svg"
        ],
        "title": "ProcessDiagram"
      },
      "ProcessUnit": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Ready"
              },
              {
                "$ref": "#/components/schemas/Error"
              },
              {
                "$ref": "#/components/schemas/InProgress"
              },
              {
                "$ref": "#/components/schemas/NotOnline"
              }
            ],
            "title": "State"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "runtime_msec": {
            "type": "integer",
            "title": "Runtime Msec"
          },
          "current_user_role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "uod_author_name": {
            "type": "string",
            "title": "Uod Author Name"
          },
          "uod_author_email": {
            "type": "string",
            "title": "Uod Author Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "state",
          "current_user_role"
        ],
        "title": "ProcessUnit",
        "description": "Represents a process unit. "
      },
...
    }
  }
}